refactor: 패키지명 변경 및 구글 컨벤션에 맞게 수정 (#243)
