+++ s3-proxy/s3-proxy/src/main/java/com/woowacourse/s3_proxy/common/file_validator/FileValidator.java
@@ -1,0 +1,8 @@
+package com.woowacourse.s3_proxy.common.file_validator;
+
+import org.springframework.web.multipart.MultipartFile;
+
+public interface FileValidator {
+
+    void execute(MultipartFile multipartFile);
+}
