+++ s3-proxy/s3-proxy/src/main/java/com/woowacourse/s3_proxy/web/presentation/resolver/ExtensionValid.java
@@ -1,0 +1,12 @@
+package com.woowacourse.s3_proxy.web.presentation.resolver;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface ExtensionValid {
+
+}
