--- s3-proxy/s3-proxy/src/main/java/com/woowacourse/s3proxy/S3ProxyApplication.java
@@ -1,13 +1,0 @@
-package com.woowacourse.s3proxy;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class S3ProxyApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(S3ProxyApplication.class, args);
-    }
-
-}
