--- s3-proxy/s3-proxy/src/main/java/com/woowacourse/s3proxy/common/filevalidator/FileValidator.java
@@ -1,8 +1,0 @@
-package com.woowacourse.s3proxy.common.filevalidator;
-
-import org.springframework.web.multipart.MultipartFile;
-
-public interface FileValidator {
-
-    void execute(MultipartFile multipartFile);
-}
