--- s3-proxy/s3-proxy/src/main/java/com/woowacourse/s3proxy/config/FileValidatorConfiguration.java
@@ -1,18 +1,0 @@
-package com.woowacourse.s3proxy.config;
-
-import com.woowacourse.s3proxy.common.filevalidator.FileValidator;
-import com.woowacourse.s3proxy.common.filevalidator.ImageFileValidator;
-import java.util.List;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class FileValidatorConfiguration {
-
-    @Bean
-    public List<FileValidator> fileValidator() {
-        return List.of(
-            new ImageFileValidator()
-        );
-    }
-}
