--- s3-proxy/s3-proxy/src/main/java/com/woowacourse/s3proxy/exception/ApplicationException.java
@@ -1,32 +1,0 @@
-package com.woowacourse.s3proxy.exception;
-
-import org.springframework.http.HttpStatus;
-
-public class ApplicationException extends RuntimeException {
-
-    private final String errorCode;
-    private final HttpStatus httpStatus;
-
-    public ApplicationException(String errorCode, HttpStatus httpStatus, String message) {
-        this(errorCode, httpStatus, message, null);
-    }
-
-    public ApplicationException(
-        String errorCode,
-        HttpStatus httpStatus,
-        String message,
-        Exception e
-    ) {
-        super(message, e);
-        this.errorCode = errorCode;
-        this.httpStatus = httpStatus;
-    }
-
-    public String getErrorCode() {
-        return errorCode;
-    }
-
-    public HttpStatus getHttpStatus() {
-        return httpStatus;
-    }
-}
