--- s3-proxy/s3-proxy/src/main/java/com/woowacourse/s3proxy/exception/format/HashFailureException.java
@@ -1,13 +1,0 @@
-package com.woowacourse.s3proxy.exception.format;
-
-import com.woowacourse.s3proxy.exception.ApplicationException;
-import org.springframework.http.HttpStatus;
-
-public class HashFailureException extends ApplicationException {
-    private static final String ERROR_CODE = "I0003";
-    private static final HttpStatus HTTP_STATUS = HttpStatus.BAD_REQUEST;
-
-    public HashFailureException(String message) {
-        super(ERROR_CODE, HTTP_STATUS, message);
-    }
-}
