+++ s3-proxy/s3-proxy/src/test/java/com/woowacourse/s3_proxy/S3ProxyApplicationTests.java
@@ -1,0 +1,13 @@
+package com.woowacourse.s3_proxy;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class S3ProxyApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
