--- s3-proxy/s3-proxy/src/test/java/com/woowacourse/s3proxy/S3ProxyApplicationTests.java
@@ -1,13 +1,0 @@
-package com.woowacourse.s3proxy;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class S3ProxyApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
